pyproject.toml
[tool.poetry]
name = "tee-baker-server"
version = "0.1.0"
description = "Backend server for tee-baker app which is used to generate t-shirt design"
authors = ["Huzaifa Arshad <huzaifa.arshad@emumba.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
psycopg = {extras = ["binary", "pool"], version = "^3.2.1"}
passlib = "^1.7.4"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
